Initial commit of new Amiga 4000 system (WIP)
New Amiga 4000 machine type, including AGA chipset emulation.
A4000 IDE support is implemented, but not 3.5" floppy emulation (yet).
The default system type is PAL, with an "ntsc" machine option boolean.
The mouse is on port 1 and there's no joystick support (yet).

The code doesn't work yet but it should at least compile.
I hope to complete the project over the next several weeks.
The following are the release notes for the planed 1.0 release.

The left-Amiga key is mapped to left-Meta (aka left-Windows key).
The right-Amiga key is mapped to right-Ctrl, right-Meta, and Menu,
to accommodate keyboards which may only have one of those keys.
The End key (and Help, if present) are mapped to Help, and the
Home and PgUp keys map to numeric keypad left and right parens.
PgDn is mapped to the international key left of Return, and both
Insert and the angle brackets key, if present, are mapped to the
angle brackets key to the right of left Shift. The optional A3000
10-second warm reset warning for Ctrl-Amiga-Amiga is implemented.
Mouse wheel events are sent through keyboard scancodes as well.

Planned features:
 - floppy disk images, including .adf and MS-DOS 720K/1.44M formats
 - optimized screen refresh mode for low-bandwidth remote connections
 - SCSI support (emulating Amiga 4000T built-in controller)
 - RS-232 serial (Paula) and parallel printer (CIAs) ports
 - Zorro II/III bus support (including INT2 and INT6 GPIO lines)
 - A2065 Ethernet support (use existing QEMU Lance Ethernet)
 - Zorro expansion RAM (up to 2GB, vs. 64MB motherboard RAM)
 - Picasso II/IV or other supported RTG-compatible graphics card
 - USB gamepad (and second mouse support if possible)
 - configurable stereo separation and Amiga low-pass audio filter
 - interface to real Amiga devices via Raspberry Pi GPIO lines
 - PCI bus interface and virtio devices (with AmigaOS support)

Three aspects of the A4000 have been intentionally left unimplemented.
The power-on overlay of the Kickstart ROM at address 0x0 in place of
chip RAM is not implemented because the CPU reset has been configured
to jump into the Kickstart ROM address directly. Kickstart immediately
turns off the OVL bit in order to access chip RAM, and it would make
zero sense to re-enable that bit and lose access to chip RAM.

The copper / blitter DMA interaction bug hasn't been implemented,
although if it would apply, a DPRINTF() warning is printed, in order
to see if it would make sense to add an option for bug-compatibility.

The last "feature" which was intentionally omitted is the A4000-only
chip register shadow in the 1MB region at 0x00c00000, where 24-bit
addressing Amigas sometimes map fast or pseudo-fast RAM. The CBM
chip specification for Gary (4393) says this shadow mapping is for
compatibility, but it makes no sense to map chip registers where
some older program may be expecting to see RAM, if anything. So the
address space from 0xc00000 to 0xcfffff is left unmapped instead.